+++ app/build.gradle
@@ -49,6 +49,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.1.0'
     implementation 'me.grantland:autofittextview:0.2.+'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
+    implementation 'com.afollestad.material-dialogs:core:3.3.0'
 
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
