+++ app/src/main/java/archiveasia/jp/co/hakenman/activity/WorksheetListActivity.kt
@@ -2,22 +2,22 @@ package archiveasia.jp.co.hakenman.activity
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.Gravity
 import android.view.Menu
 import android.view.MenuItem
-import android.widget.TextView
-import androidx.appcompat.app.AlertDialog
+import android.view.View
+import android.widget.NumberPicker
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import archiveasia.jp.co.hakenman.CreateWorksheetDialog
 import archiveasia.jp.co.hakenman.CustomLog
 import archiveasia.jp.co.hakenman.R
 import archiveasia.jp.co.hakenman.adapter.WorksheetListAdapter
 import archiveasia.jp.co.hakenman.manager.WorksheetManager
 import archiveasia.jp.co.hakenman.model.Worksheet
+import com.afollestad.materialdialogs.MaterialDialog
+import com.afollestad.materialdialogs.customview.customView
+import com.afollestad.materialdialogs.customview.getCustomView
 import kotlinx.android.synthetic.main.activity_main.*
-import java.text.SimpleDateFormat
-import java.util.Locale
+import kotlinx.android.synthetic.main.datepicker_dialog.view.*
 
 class WorksheetListActivity : AppCompatActivity() {
 
@@ -36,8 +36,12 @@ class WorksheetListActivity : AppCompatActivity() {
             }
 
             override fun onLongClickItem(index: Int) {
-                showAlertDialog(getString(R.string.delete_worksheet_title), getString(R.string.delete_button)) {
-                    adapter.replaceWorksheetList(WorksheetManager.removeWorksheet(index))
+                MaterialDialog(this@WorksheetListActivity).show {
+                    message(R.string.delete_worksheet_title)
+                    positiveButton(R.string.delete_button) {
+                        adapter.replaceWorksheetList(WorksheetManager.removeWorksheet(index))
+                    }
+                    negativeButton(R.string.negative_button)
                 }
             }
 
@@ -72,43 +76,46 @@ class WorksheetListActivity : AppCompatActivity() {
         return super.onOptionsItemSelected(item)
     }
 
-    private fun showAlertDialog(title: String, btn: String, completion: () -> Unit) {
-        val alertDialog = AlertDialog.Builder(this)
-        with (alertDialog) {
-            val titleView = TextView(context)
-            titleView.text = title
-            titleView.gravity = Gravity.CENTER_HORIZONTAL
-            titleView.textSize = 20F
-            titleView.setTextColor(resources.getColor(R.color.colorBlack))
-            setView(titleView)
-
-            setPositiveButton(btn) {
-                dialog, whichButton ->
-                completion()
-            }
-
-            setNegativeButton(getString(R.string.negative_button)) {
-                dialog, whichButton ->
-                dialog.dismiss()
-            }
-        }
-        val dialog = alertDialog.create()
-        dialog.show()
-    }
-
-    private fun showCreateWorksheetDialog() {
-        val dialog = CreateWorksheetDialog(object : CreateWorksheetDialog.WorksheetDialogListener{
-            override fun createdWorksheet() {
-                CustomLog.d("createdWorksheet")
-                reloadListView()
-            }
-        })
-        dialog.show(supportFragmentManager, "tag")
-        supportFragmentManager.executePendingTransactions()
-    }
-
     private fun reloadListView() {
         val worksheetList = WorksheetManager.getWorksheetList()
         adapter.replaceWorksheetList(worksheetList)
+    }
+
+    private fun showCreateWorksheetDialog() {
+        val dialog = MaterialDialog(this).customView(R.layout.datepicker_dialog)
+        dialog.title(R.string.create_worksheet_title)
+        val customView = dialog.getCustomView()
+        val datePicker = customView.date_picker.apply {
+            // 日のSpinnerを非表示する
+            findViewById<NumberPicker>(resources.getIdentifier("day", "id", "android")).visibility = View.GONE
+            // 当月まで選択できるように
+            maxDate = System.currentTimeMillis()
+        }
+        dialog.positiveButton(R.string.positive_button) {
+            val year = datePicker.year.toString()
+            var month = (datePicker.month + 1).toString()
+            if (month.length < 2) {
+                month = "0$month"
+            }
+            val yearMonth = year + month
+            val worksheet = WorksheetManager.createWorksheet(yearMonth)
+            if (WorksheetManager.isAlreadyExistWorksheet(yearMonth)) {
+                MaterialDialog(this).show {
+                    message(R.string.update_worksheet_title)
+                    positiveButton(R.string.positive_button) {
+                        WorksheetManager.updateWorksheet(worksheet)
+                        CustomLog.d("勤務表生成 : $yearMonth")
+                        reloadListView()
+                    }
+                    negativeButton(R.string.negative_button)
+                }
+            } else {
+                WorksheetManager.addWorksheetToJsonFile(worksheet)
+                CustomLog.d("勤務表生成 : $yearMonth")
+                reloadListView()
+            }
+        }
+        dialog.negativeButton(R.string.negative_button)
+        dialog.show()
     }
 }
