+++ app/src/main/res/layout/datepicker_dialog.xml
@@ -1,55 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <TextView
-        android:id="@+id/text_dialog_header"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:padding="@dimen/large_padding"
-        android:text="@string/create_worksheet_title"
-        android:textSize="@dimen/large_font_size"
-        android:textColor="@color/colorWhite"
-        android:background="@color/colorPrimary"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content">
 
     <DatePicker
         android:id="@+id/date_picker"
-        android:layout_width="254dp"
-        android:layout_height="234dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginEnd="32dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:calendarViewShown="false"
         android:datePickerMode="spinner"
+        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/text_dialog_header" />
-
-    <Button
-        android:id="@+id/button_negative"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:text="@string/negative_button"
-        android:backgroundTint="@color/colorLightGray"
-        app:layout_constraintTop_toBottomOf="@id/date_picker"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/button_positive"
-        app:layout_constraintBottom_toBottomOf="parent"/>
-
-    <Button
-        android:id="@+id/button_positive"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:text="@string/positive_button"
-        android:textColor="@color/colorWhite"
-        android:backgroundTint="@color/colorPrimary"
-        app:layout_constraintTop_toBottomOf="@id/date_picker"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/button_negative"
-        app:layout_constraintBottom_toBottomOf="parent"/>
+        app:layout_constraintBottom_toBottomOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
